New logic for furin cleavage site annotation.
This updated script is mostly inspired by the quickstart build, but
with one small change. It was discovered that sequences being out
of frame could result in mistranslations and misannotations. The
script now scans the alignment for the earliest occurrence of the
start codon, and uses this as a basis for translation.